Reorganise entity/repository packages
- Add RedisURL entity
- Relocate URL entity/repository
