+++ src/main/java/com/elliot/tinyurl/repository/redis/RedisURLRepository.java
@@ -1,0 +1,7 @@
+package com.elliot.tinyurl.repository.redis;
+
+import com.elliot.tinyurl.entity.redis.RedisURL;
+import org.springframework.data.jpa.repository.JpaRepository;
+
+public interface RedisURLRepository extends JpaRepository<RedisURL, String> {
+}
